requirements.txt
pyyaml==6.0.1
requests==2.32.3
spotipy==2.22.1
ytmusicapi==1.7.3